requirements.txt
hyperliquid-python-sdk>=0.7.0
eth-account>=0.13
anthropic>=0.40.0
discord-webhook>=1.3.0
feedparser>=6.0.10
aiohttp>=3.9.0
pydantic>=2.5.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
